remove unnecessary nil check
